Created and implemented Preferences system
Preferences manages all preferency things. It loads in before the rest of the app starts
Preferences is static so the preferences and it's functions can be accessed from anywere
Settings got a better opening/ closing animation
Settings now implements Preferences and all currently existing preference values
Settings is now styled "better". The entire app needs some styling and a color scheme
SettingsItem is a Widget that is meant to contain a preference. It holds a title and a widget that modifies the preference